Make the generated dll file all in one.

+++ src/CDMWrapper/CDMWrapper.cs
@@ -5,9 +5,9 @@ namespace CDMWrapper
 {
     public class CDMWrapper
     {
-        public void showCDM(long param)
+        public void showCDM(IntPtr param)
         {
-            IntPtr hwnd = (IntPtr)param;
+            IntPtr hwnd = param;
             System.Windows.Interop.HwndSourceParameters sourceParams = new System.Windows.Interop.HwndSourceParameters("CDMWrapper");
             sourceParams.ParentWindow = hwnd;
             sourceParams.WindowStyle = 0x10000000 | 0x40000000; // WS_VISIBLE | WS_CHILD; // style
